requirements.txt
langchain-groq
chromadb
langchain_community 
beautifulsoup4
streamlit
python-dotenv